backend/requirements.txt
Flask==3.0.0
SpeechRecognition==3.15.1
requests==2.31.0
python-dotenv==1.0.0
resemblyzer>=0.1.4
numpy>=1.20.0
soundfile>=0.12.1
